add testcase and refactor
