+++ e2e/testcases/birth/declarations/community-leader-notifies-birth.spec.ts
@@ -1,53 +1,65 @@
-import { test, expect, type Page } from '@playwright/test'
-import { formatName, goToSection, login } from '../../../helpers'
+import { test, expect } from '@playwright/test'
+import {
+  drawSignature,
+  formatName,
+  goToSection,
+  login,
+  selectDeclarationAction
+} from '../../../helpers'
 import { faker } from '@faker-js/faker'
 import { CREDENTIALS } from '../../../constants'
 import { ensureOutboxIsEmpty } from '../../../utils'
-import { selectDeclarationAction } from '../../../helpers'
 
-test.describe.serial('Community leader notifies birth', () => {
-  let page: Page
-
+test('Community leader notifies birth', async ({ page }) => {
   const childName = {
     firstNames: faker.person.firstName(),
     familyName: faker.person.lastName()
   }
 
-  test.beforeAll(async ({ browser }) => {
-    page = await browser.newPage()
-  })
-
-  test.afterAll(async () => {
-    await page.close()
-  })
-
-  test('Login as Community leader', async () => {
+  await test.step('Login as Community leader', async () => {
     await login(page, CREDENTIALS.COMMUNITY_LEADER)
   })
 
-  test('Initiate birth declaration', async () => {
+  await test.step('Initiate birth declaration', async () => {
     await page.click('#header-new-event')
     await page.getByLabel('Birth').click()
     await page.getByRole('button', { name: 'Continue' }).click()
     await page.getByRole('button', { name: 'Continue' }).click()
   })
 
-  test('Fill child details', async () => {
+  await test.step('Fill child details', async () => {
     await page.locator('#firstname').fill(childName.firstNames)
     await page.locator('#surname').fill(childName.familyName)
   })
 
-  test('Continue to review', async () => {
+  await test.step('Continue to review', async () => {
     await goToSection(page, 'review')
   })
 
-  test('Notify', async () => {
+  await test.step('Provide informant signature, then delete it', async () => {
+    await page.getByRole('button', { name: 'Sign', exact: true }).click()
+    await drawSignature(page, 'review____signature_canvas_element', false)
+    await page
+      .locator('#review____signature_modal')
+      .getByRole('button', { name: 'Apply' })
+      .click()
+
+    await expect(page.getByRole('dialog')).not.toBeVisible()
+
+    await page.getByRole('button', { name: 'Delete', exact: true }).click()
+
+    await expect(
+      page.getByRole('button', { name: 'Sign', exact: true })
+    ).toBeVisible()
+  })
+
+  await test.step('Notify', async () => {
     await selectDeclarationAction(page, 'Notify')
 
     await ensureOutboxIsEmpty(page)
   })
 
-  test('Open record', async () => {
+  await test.step('Open record', async () => {
     await page.getByText('Recent').click()
 
     await page
@@ -57,7 +69,7 @@ test.describe.serial('Community leader notifies birth', () => {
       .click()
   })
 
-  test('Assert that record is notified', async () => {
+  await test.step('Assert that record is notified', async () => {
     await expect(page.getByTestId('status-value')).toHaveText('Notified')
   })
 })
